requirements.txt
Django==2.2.12
djangorestframework==3.11.0
django-rest-swagger====2.2.0
